+++ changelog.md
@@ -10,6 +10,7 @@ API changes:
   * The predefined task types have been renamed from "REGULAR" to "COMPUTATIONAL" and from "IO" to "BLOCKING" (motivated by the post [Be Aware of ForkJoinPool#commonPool()](https://dzone.com/articles/be-aware-of-forkjoinpoolcommonpool])
   * The method `submit()` has been removed from the interface `Handle` because it is for internal use only.
   * The method `join()` has been removed from the interface `Handle` because the idea behind Hippodamus discourages explicit joins. Instead, users should specify the handle a task should join on as the task's dependency.
+  * The methods `stop()` and `hasStopped()` have been removed from the interface `Handle` because in the context of Hippodamus it does not make sense to stop some but not all tasks. Use the method `ExecutionCoordinator.stop()` to stop all tasks.
   * The interfaces `StoppableExceptionalRunnable` and `StoppableExceptionalCallable` have been removed. A task now simply has to check `Thread.isInterrupted()` or `Thread.interrupted()` to determine whether it should stop instead of checking the Boolean supplier `stopFlag`.
   * The enum `WaitMode` has been removed. You cannot specify anymore how long the `ExecutionCoordinator` waits when all tasks have terminated or been stopped. Now, the coordinator's `close()` method will not return as long as any of its tasks is being executed. This behavior has formerly been described by `WaitMode.UNTIL_TERMINATION`.